Add notifyapi package containing Error
